feat: 예시 프로필 이미지 API 요청을 커스텀 훅으로 분리
