+++ src/pages/message/index.jsx
@@ -1,17 +1,15 @@
 import InputWithError from './components/InputWithError';
 import ProfileImg from './components/ProfileImg';
-import api from '../../api/axios';
 import profilePreview from '../../assets/images/profile.svg';
 import Select from './components/Select';
 import Editor from './components/Editor';
 import Button from '../../components/common/button/index';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
+import useProfileImages from './utils/useProfileImages';
 
 const Message = () => {
   const [sender, setSender] = useState('');
   const [isValid, setIsValid] = useState(true);
-  const [images, setImages] = useState([]);
-  const [loadingError, setLoadingError] = useState(false);
   const [profileImg, setProfileImg] = useState(profilePreview);
   const [selectedRelation, setSelectedRelation] = useState('지인');
   const [editorContent, setEditorContent] = useState('');
@@ -20,24 +18,11 @@ const Message = () => {
   const fontOptions = ['Noto Sans', 'Pretendard', '나눔명조', '나눔손글씨 손편지체'];
 
   const handleInputChange = (e) => {
-    setSender(e.target.value); // 공백 제거 후 상태 업데이트
+    setSender(e.target.value.trim()); // 공백 제거 후 상태 업데이트
     setIsValid(true);
   };
 
-  // 예시 프로필 이미지 요청
-  useEffect(() => {
-    const getProfileImages = async () => {
-      try {
-        const response = await api.getProfileImages();
-        setImages(response.data.imageUrls);
-        setLoadingError(false);
-      } catch (e) {
-        setLoadingError(true);
-        console.error(e);
-      }
-    };
-    getProfileImages();
-  }, []);
+  const { images, loadingError } = useProfileImages();
 
   return (
     <form className="container max-w-[720px] mx-auto mt-[47px] flex flex-col ">
